+++ returns/curry.py
@@ -1,14 +1,14 @@
-from functools import partial
+from functools import partial as _partial
 from typing import Any, Callable, TypeVar
 
 _ReturnType = TypeVar('_ReturnType')
 
 
-def curry(
+def partial(
     func: Callable[..., _ReturnType], *args: Any, **kwargs: Any,
 ) -> Callable[..., _ReturnType]:
     """
-    Typed curring helper.
+    Typed partial application helper.
 
     It just ``functools.partial`` wrapper with better typing support.
 
@@ -17,12 +17,17 @@ def curry(
 
     .. code:: python
 
-        >>> from returns.curry import curry
-        >>> curried_int = curry(int, base=2)
-        >>> assert curried_int('10') == 2
+        >>> from returns.curry import partial
+
+        >>> def sum_two_numbers(first: int, second: int) -> int:
+        ...     return first + second
+        ...
+        >>> sum_with_ten = partial(sum_two_numbers, 10)
+        >>> assert sum_with_ten(2) == 12
+        >>> assert sum_with_ten(-5) == 5
 
     See also:
         https://docs.python.org/3/library/functools.html#functools.partial
 
     """
-    return partial(func, *args, **kwargs)
+    return _partial(func, *args, **kwargs)